+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## Unreleased
+
+- Add distinct exit regions to watershed process
+- Expose direct border tiles in `RoomRegionAnalysis` via `get_border_tiles` function
+
+## 0.1.2
+
+- Improve documentation
+
 ## 0.1.1
 
 - Add `get_region_analysis_for_room_by_terrain` function